requirements.txt
pytest==7.1.3
selene==2.0.0b12
allure-pytest